Correccion alta producto

+++ src/main/java/com/ipn/mx/modelo/entidades/Producto.java
@@ -6,6 +6,7 @@
 package com.ipn.mx.modelo.entidades;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 //import lombok.Data;
 //import lombok.NoArgsConstructor;
 
@@ -18,7 +19,7 @@ public class Producto implements Serializable{
     private int idProducto;
     private String nombreProducto;
     private String descripcionProducto;
-    private double precio;
+    private BigDecimal precio;
     private int existencia;
     private int stockMinimo;
     private int claveCategoria;
@@ -50,11 +51,11 @@ public class Producto implements Serializable{
         this.descripcionProducto = descripcionProducto;
     }
 
-    public double getPrecio() {
+    public BigDecimal getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(BigDecimal precio) {
         this.precio = precio;
     }
